extract CanvasSidebarToolbar from CanvasPage
